Add files via upload
Quantum computing, a revolutionary paradigm in computation, harnesses the principles of quantum mechanics to perform computations that were once thought to be impossible for classical computers. Grover's algorithm, a prominent quantum algorithm, stands out for its potential to significantly speed up unstructured search problems. This paper explores the application of artificial intelligence (AI) techniques in the understanding and optimization of Grover's algorithm.

Quantum computing and artificial intelligence (AI) are two cutting-edge technologies that have the potential to revolutionize various fields. While they are distinct fields, the intersection of quantum computing and AI holds promise for solving complex problems more efficiently than classical methods.

while quantum computing is still in its early stages of development, it has the potential to impact various aspects of artificial intelligence. Researchers and practitioners are actively working on harnessing the unique capabilities of quantum computers to enhance AI algorithms and solve problems that are currently intractable for classical computers.